Claude has refactored all LLM providers
The Anthropic streaming test currently fails
with an additional empty chunk in the beginning.
