+++ src/llm/anthropic/rate_limits.rs
@@ -1,0 +1,103 @@
+
+use crate::llm::types::RateLimitHandler;
+use chrono::{DateTime, Utc};
+use reqwest::Response;
+use std::time::Duration;
+use tracing::debug;
+
+#[derive(Debug)]
+pub struct AnthropicRateLimitInfo {
+    requests_limit: Option<u32>,
+    requests_remaining: Option<u32>,
+    requests_reset: Option<DateTime<Utc>>,
+    tokens_limit: Option<u32>,
+    tokens_remaining: Option<u32>,
+    tokens_reset: Option<DateTime<Utc>>,
+    retry_after: Option<Duration>,
+}
+
+impl RateLimitHandler for AnthropicRateLimitInfo {
+    fn from_response(response: &Response) -> Self {
+        let headers = response.headers();
+
+        fn parse_header<T: std::str::FromStr>(
+            headers: &reqwest::header::HeaderMap,
+            name: &str,
+        ) -> Option<T> {
+            headers
+                .get(name)
+                .and_then(|h| h.to_str().ok())
+                .and_then(|s| s.parse().ok())
+        }
+
+        fn parse_datetime(
+            headers: &reqwest::header::HeaderMap,
+            name: &str,
+        ) -> Option<DateTime<Utc>> {
+            headers
+                .get(name)
+                .and_then(|h| h.to_str().ok())
+                .and_then(|s| DateTime::parse_from_rfc3339(s).ok())
+                .map(|dt| dt.into())
+        }
+
+        Self {
+            requests_limit: parse_header(headers, "anthropic-ratelimit-requests-limit"),
+            requests_remaining: parse_header(headers, "anthropic-ratelimit-requests-remaining"),
+            requests_reset: parse_datetime(headers, "anthropic-ratelimit-requests-reset"),
+            tokens_limit: parse_header(headers, "anthropic-ratelimit-tokens-limit"),
+            tokens_remaining: parse_header(headers, "anthropic-ratelimit-tokens-remaining"),
+            tokens_reset: parse_datetime(headers, "anthropic-ratelimit-tokens-reset"),
+            retry_after: parse_header::<u64>(headers, "retry-after").map(Duration::from_secs),
+        }
+    }
+
+    fn get_retry_delay(&self) -> Duration {
+        // If we have a specific retry-after duration, use that
+        if let Some(retry_after) = self.retry_after {
+            return retry_after;
+        }
+
+        // Otherwise, calculate based on reset times
+        let now = Utc::now();
+        let mut shortest_wait = Duration::from_secs(60); // Default to 60 seconds if no information
+
+        // Check requests reset time
+        if let Some(reset_time) = self.requests_reset {
+            if reset_time > now {
+                shortest_wait = shortest_wait.min(Duration::from_secs(
+                    (reset_time - now).num_seconds().max(0) as u64,
+                ));
+            }
+        }
+
+        // Check tokens reset time
+        if let Some(reset_time) = self.tokens_reset {
+            if reset_time > now {
+                shortest_wait = shortest_wait.min(Duration::from_secs(
+                    (reset_time - now).num_seconds().max(0) as u64,
+                ));
+            }
+        }
+
+        // Add a small buffer to avoid hitting the limit exactly at reset time
+        shortest_wait + Duration::from_secs(1)
+    }
+
+    fn log_status(&self) {
+        debug!(
+            "Rate limits - Requests: {}/{} (reset: {}), Tokens: {}/{} (reset: {})",
+            self.requests_remaining
+                .map_or("?".to_string(), |r| r.to_string()),
+            self.requests_limit
+                .map_or("?".to_string(), |l| l.to_string()),
+            self.requests_reset
+                .map_or("unknown".to_string(), |r| r.to_string()),
+            self.tokens_remaining
+                .map_or("?".to_string(), |r| r.to_string()),
+            self.tokens_limit.map_or("?".to_string(), |l| l.to_string()),
+            self.tokens_reset
+                .map_or("unknown".to_string(), |r| r.to_string()),
+        );
+    }
+}
