+++ src/llm/vertex/rate_limits.rs
@@ -1,0 +1,37 @@
+
+use crate::llm::types::RateLimitHandler;
+use reqwest::Response;
+use std::time::Duration;
+use tracing::debug;
+
+/// Rate limit information extracted from response headers
+#[derive(Debug)]
+pub struct VertexRateLimitInfo {
+    // TODO: Add actual rate limit fields once we know what headers Vertex AI uses
+    requests_remaining: Option<u32>,
+    #[allow(dead_code)]
+    requests_reset: Option<Duration>,
+}
+
+impl RateLimitHandler for VertexRateLimitInfo {
+    fn from_response(_response: &Response) -> Self {
+        // TODO: Parse actual rate limit headers once we know what Vertex AI provides
+        Self {
+            requests_remaining: None,
+            requests_reset: None,
+        }
+    }
+
+    fn get_retry_delay(&self) -> Duration {
+        // Default exponential backoff strategy
+        Duration::from_secs(2)
+    }
+
+    fn log_status(&self) {
+        debug!(
+            "Vertex AI Rate limits - Requests remaining: {}",
+            self.requests_remaining
+                .map_or("unknown".to_string(), |r| r.to_string())
+        );
+    }
+}
